errorprone: fix errors
* https://errorprone.info/bugpattern/BadShiftAmount
* https://errorprone.info/bugpattern/EqualsHashCode
* https://errorprone.info/bugpattern/ChainingConstructorIgnoresParameter

+++ hollow/src/main/java/com/netflix/hollow/tools/patch/delta/HollowStateDeltaPatcher.java
@@ -418,11 +418,11 @@ public class HollowStateDeltaPatcher {
                 
                 HollowMapEntryOrdinalIterator iter = fromMapState.ordinalIterator(ordinal);
                 while(iter.next())
-                    fromScratch.add((iter.getKey() << 32) | iter.getValue());
+                    fromScratch.add(((long)iter.getKey() << 32) | iter.getValue());
                 
                 iter = toMapState.ordinalIterator(ordinal);
                 while(iter.next())
-                    toScratch.add((iter.getKey() << 32) | iter.getValue());
+                    toScratch.add(((long)iter.getKey() << 32) | iter.getValue());
                 
                 fromScratch.sort();
                 toScratch.sort();
